Title: 합과 평균, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240801/합과 평균/sum-and-avg.cpp
@@ -9,7 +9,7 @@ int main() {
     c = a + b;
 
     int d;
-    d = (a + b) / 2;
+    d = (double) (a + b) / 2;
 
     cin >> a >> b;
     cout << c << " " << d;
